+++ src/main/java/com/team7/retriever/domain/agrot/domain/repository/ArgotsRepository.java
@@ -1,4 +1,4 @@
-package com.team7.retriever.repository;
+package com.team7.retriever.domain.agrot.domain.repository;
 
 import com.team7.retriever.domain.agrot.domain.document.Argot;
