+++ src/main/java/com/team7/retriever/domain/chatbot/domain/repository/ChatBotsRepository.java
@@ -1,4 +1,4 @@
-package com.team7.retriever.repository;
+package com.team7.retriever.domain.chatbot.domain.repository;
 
 import com.team7.retriever.domain.chatbot.domain.document.ChatBots;
 import org.springframework.data.mongodb.repository.MongoRepository;
