+++ src/main/java/com/team7/retriever/domain/crawling/domain/repository/PostHtmlRepository.java
@@ -1,4 +1,4 @@
-package com.team7.retriever.repository;
+package com.team7.retriever.domain.crawling.domain.repository;
 
 import com.team7.retriever.domain.crawling.domain.document.PostHtml;
 import org.springframework.data.mongodb.repository.MongoRepository;
