+++ src/main/java/com/team7/retriever/domain/drug/domain/repository/DrugsRepository.java
@@ -1,4 +1,4 @@
-package com.team7.retriever.repository;
+package com.team7.retriever.domain.drug.domain.repository;
 
 import com.team7.retriever.domain.drug.domain.document.Drugs;
