+++ src/main/java/com/team7/retriever/domain/report/domain/repository/ReportsRepository.java
@@ -1,4 +1,4 @@
-package com.team7.retriever.repository;
+package com.team7.retriever.domain.report.domain.repository;
 
 import com.team7.retriever.domain.report.domain.document.Reports;
 import org.springframework.data.mongodb.repository.MongoRepository;
