+++ src/main/java/com/team7/retriever/service/ArgotsService.java
@@ -1,7 +1,7 @@
 package com.team7.retriever.service;
 
 import com.team7.retriever.domain.agrot.domain.document.Argot;
-import com.team7.retriever.repository.ArgotsRepository;
+import com.team7.retriever.domain.agrot.domain.repository.ArgotsRepository;
 
 import org.springframework.stereotype.Service;
