+++ src/main/java/com/team7/retriever/service/ChInfoService.java
@@ -1,7 +1,7 @@
 package com.team7.retriever.service;
 
 import com.team7.retriever.domain.channel.domain.document.ChInfo;
-import com.team7.retriever.repository.ChInfoRepository;
+import com.team7.retriever.domain.channel.domain.repository.ChInfoRepository;
 
 import org.springframework.stereotype.Service;
