+++ src/main/java/com/team7/retriever/service/PostsService.java
@@ -2,7 +2,7 @@ package com.team7.retriever.service;
 
 import com.team7.retriever.domain.post.controller.dto.request.PostUpdateCheckRequest;
 import com.team7.retriever.domain.post.domain.document.Posts;
-import com.team7.retriever.repository.PostsRepository;
+import com.team7.retriever.domain.post.domain.repository.PostsRepository;
 
 import org.springframework.stereotype.Service;
 import org.springframework.data.domain.Sort;
