+++ src/main/java/com/team7/retriever/service/PreprocessService.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.team7.retriever.domain.crawling.controller.dto.response.PreprocessResponse;
 import com.team7.retriever.domain.post.domain.document.Posts;
-import com.team7.retriever.repository.PostsRepository;
+import com.team7.retriever.domain.post.domain.repository.PostsRepository;
 
 import lombok.RequiredArgsConstructor;
