+++ src/main/java/com/team7/retriever/service/ReportsService.java
@@ -1,7 +1,7 @@
 package com.team7.retriever.service;
 
 import com.team7.retriever.domain.report.domain.document.Reports;
-import com.team7.retriever.repository.ReportsRepository;
+import com.team7.retriever.domain.report.domain.repository.ReportsRepository;
 
 import org.springframework.stereotype.Service;
